requirements.txt
streamlit
unstructured
pdf2image
pandas
langchain
chromadb
google-generativeai
langchain-google-genai
python-dotenv
tavily-python